Fetch Amazon products only when tab is opened
